chore: fix recent regression with `deno upgrade` not handling redirects (#17045)
